+++ lib/message_processor/message_processor.rb
@@ -5,23 +5,14 @@ class MessageProcessor
   end
 
   def collectUserInfo(data:)
-    user = User.where(slack_user_id: data.user).first
-    if user.blank?
-      user_info = @slack_web_client.users_info(user: data.user)
-
-      user = User.new
-      user.slack_user_id = user_info["user"]["id"]
+    user_info = @slack_web_client.users_info(user: data.user)
+    user = User.find_or_initialize_by(slack_user_id: user_info["user"]["id"])
+    if user.new_record? || user.updated_at < 1.day.ago
       user.name = user_info["user"]["name"]
       user.real_name = user_info["user"]["real_name"]
       user.tz = user_info["user"]["tz"]
       user.tz_offset = user_info["user"]["tz_offset"]
       user.contacts = user_info["user"]["profile"]["email"]
-      user.save
-    elsif user.updated_at < 1.day.ago
-      user_info = @slack_web_client.users_info(user: data.user)
-      user.tz = user_info["user"]["tz"]
-      user.tz_offset = user_info["user"]["tz_offset"]
-      user.updated_at = 1.minutes.ago
       user.save
     end
   end